git-svn-id: https://svn.kuali.org/repos/student/poc/brms/brms-ri/branches/brms-ri-dev@789 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ generate-rule/src/main/java/org/kuali/student/rules/statement/SubsetProposition.java
@@ -69,7 +69,7 @@ public class SubsetProposition<T extends Integer, E> extends AbstractProposition
 
         // TODO: Use the operator to compute exact message
         Set<E> unMet = andNot();
-        int needed = (Integer) expectedValue - count;
+        int needed = expectedValue - count;
         String advice = String.format(format, needed, unMet.toString());
         report.setFailureMessage(advice);
     }
